keymanagement/requirements.txt
fastapi
uvicorn
requests
pytest
openai
python-dotenv
azure-identity
azure-keyvault-secrets
azure-storage-blob
starlette
pyyaml
